TL;DR: triton_utils.config_of determines divisibility by 16 for each of the inputs to the kernel (pointer alignment for pointers, and divisibility by 16 for sizes). For sizes, the check previously could only return true if the expr representing the size was an integer. However, it's possible for non-integral exprs to be divisible by 16, e.g. for an expr like 16*s0.

Motivation: Knowledge about divisibility by 16 allows for vectorizing loads and stores, which can improve memory bandwidth. If we have, for example, kernels with shape [s0, 16] (dynamic batch size; static, divisible-by-16 other dimensions), we want to still be able to vectorize those loads and stores.

Dashboard results suggest that this improves dynamic shape training performance for timm, and possibly a small improvement for torchbench as well. More details are provided in a comment below.

cc voznesenskym penguinwu EikanWang jgong5 Guobing-Chen XiaobingSuper zhuhaozhe blzheng Xia-Weiwen wenzhe-nrv jiayisunx peterbell10 ipiszy ngimel yf225 chenyang78 kadeng muchulee8 aakhundov

[ghstack-poisoned]

+++ test/inductor/test_codegen_triton.py
@@ -35,7 +35,7 @@ class TestCodegenTriton(TorchTestCase):
 
     @inductor_config.patch("triton.divisible_by_16", True)
     def test_config_of_sizearg(self):
-        two = sympy.Integer(4)
+        two = sympy.Integer(2)
         eight = sympy.Integer(8)
         sixteen = sympy.Integer(16)
         s0 = sympy.Symbol("s0", positive=True, integer=True)
@@ -45,11 +45,11 @@ class TestCodegenTriton(TorchTestCase):
             (2,),
             triton_utils.config_of(
                 [
-                    SizeArg("A", two),
-                    SizeArg("B", eight),
-                    SizeArg("C", sixteen),
-                    SizeArg("D", s0),
-                    SizeArg("E", s1),
+                    SizeArg("A", two),  # no
+                    SizeArg("B", eight),  # no
+                    SizeArg("C", sixteen),  # yes
+                    SizeArg("D", s0),  # no
+                    SizeArg("E", s1),  # no
                 ]
             ).divisible_by_16,
         )
